All changes

+++ src/org/j2os/monitor/modules/device/controller/DeviceGroupController.java
@@ -6,10 +6,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.*;
+
+import java.lang.reflect.InvocationTargetException;
+import java.util.List;
 
 @Controller
 @RequestMapping("/admin/deviceGroup")
@@ -38,26 +38,33 @@ public class DeviceGroupController {
         return "redirect:/admin/deviceGroup/index.do";
     }
 
-    @RequestMapping(value = "/edit/{id}")
-    public String editForm(@PathVariable long id) {
+
+    @RequestMapping(value = "/{id}/edit.do", method = RequestMethod.GET)
+    public String editForm(@PathVariable("id") long id, Model model) {
         DeviceGroup deviceGroup = (DeviceGroup) this.deviceGroupService.findById(id);
-        return "admin/admin/create";
+        model.addAttribute("devGroup", deviceGroup);
+        return "/admin/device/deviceGroupEdit";
     }
 
-    @RequestMapping(value = "/delete/{id}")
-    public String delete(@PathVariable long id) {
+    @RequestMapping(value = "/update.do", method = RequestMethod.PUT)
+    public String update(@ModelAttribute("deviceGroup") DeviceGroup deviceGroup) {
+        try {
+            deviceGroupService.update(deviceGroup);
+        } catch (InvocationTargetException e) {
+            e.printStackTrace();
+        } catch (IllegalAccessException e) {
+            e.printStackTrace();
+        }
+
+        return "redirect:/admin/deviceGroup/index.do";
+    }
+
+    @RequestMapping(value = "/delete.do", method = RequestMethod.DELETE)
+    public String delete(@RequestParam("deviceGroupId") long id) {
         DeviceGroup deviceGroup = (DeviceGroup) this.deviceGroupService.findById(id);
-        this.deviceGroupService.delete(deviceGroup);
-        return "redirect:admin/admin/";
-    }
-
-    @RequestMapping(value = "/admin/{id}", method = RequestMethod.GET)
-    public String findOne(@PathVariable long id) {
-        return "admin/admin/create";
-    }
-
-    @RequestMapping(value = "/admin/", method = RequestMethod.GET)
-    public String findAll() {
-        return "admin/admin/list";
+        if (deviceGroup != null) {
+            this.deviceGroupService.delete(deviceGroup);
+        }
+        return "redirect:/admin/deviceGroup/index.do";
     }
 }
